Improved output of exception.
Exceptions in modern Javascript have to be thrown by `new Error()` or derived types. Also QJSEngine recognizes other error types. This group of errors or exceptions are transported in a QJSValue which has isError() == true and errorType() != NoError. However exceptions also can be thrown by calling `throw 'this is an error'` or with any other literal. This kind of exceptions is transfered as is, but it is not an JS Error object.